refactor type/utils, clean up its usage
