+++ src/metricManager.ts
@@ -1,4 +1,4 @@
-import type { ActivityHeatmapData, ActivityOverTimeData, CheckpointData, DateMetrics, FileMetrics, HeatmapActivityData, MetricType } from './types'
+import type { ActivityOverTimeData, DateActivityMetrics, FileCheckpointMetrics, MetricType } from './types'
 import type ActivityHeatmapPlugin from './main'
 import type { TFile } from 'obsidian';
 import { calculateAbsoluteDifference, getDateStringFromTimestamp } from './utils';
@@ -51,14 +51,22 @@ export class MetricManager {
         return content.split(/\s+/).length;
     }
 
+
+    /**
+     * Calculates the metrics for a single file and returns the new checkpoint metrics and updated activity over time data.
+     * @param file - The Obsidian TFile to calculate metrics for
+     * @param fileCheckpointMetrics - The existing checkpoint metrics for the file, if any
+     * @param activityOverTime - The existing activity over time data (may be empty)
+     * @param isFirstTime - Whether this is a first-time user (i.e. no existing data.json file)
+     */
     async calculateFileMetrics(
         file: TFile, 
-        fileCheckpointMetrics: FileMetrics | null, 
+        fileCheckpointMetrics: FileCheckpointMetrics | null, 
         activityOverTime: ActivityOverTimeData, 
         isFirstTime: boolean
-    ): Promise<{ newFileCheckpointMetrics: FileMetrics; activityOverTime: ActivityOverTimeData}> {
+    ): Promise<{ newFileCheckpointMetrics: FileCheckpointMetrics; activityOverTime: ActivityOverTimeData}> {
         
-        const newFileCheckpointMetrics = {} as FileMetrics;
+        const newFileCheckpointMetrics = {} as FileCheckpointMetrics;
         newFileCheckpointMetrics.mtime = file.stat.mtime;
 
         for (const metricType of METRIC_TYPES) {
@@ -73,7 +81,7 @@ export class MetricManager {
                 
                 // Initialize the date object if it doesn't exist
                 if (!activityOverTime[dateString]) {
-                    activityOverTime[dateString] = {} as DateMetrics;
+                    activityOverTime[dateString] = {} as DateActivityMetrics;
                 }
                 
                 // Initialize the metric if it doesn't exist
@@ -81,12 +89,9 @@ export class MetricManager {
                     activityOverTime[dateString][metricType] = 0;
                 }
 
-                let absoluteDifference = metricValue;
-                if (fileCheckpointMetrics) {
-                    const previousValue = fileCheckpointMetrics[metricType];
-                    absoluteDifference = calculateAbsoluteDifference(metricValue, previousValue);
-                }
-                
+                const previousValue = fileCheckpointMetrics?.[metricType];
+                const absoluteDifference = calculateAbsoluteDifference(metricValue, previousValue);
+                console.log("Updating activity over time for", metricType, "by", absoluteDifference, "for date", dateString);
                 activityOverTime[dateString][metricType] += absoluteDifference;
             }
         }
@@ -94,50 +99,5 @@ export class MetricManager {
         return { newFileCheckpointMetrics, activityOverTime };
     }
 
-    /**
-     * Calculates metrics for a single file and updates the activity data.
-     * This method handles the calculation of a specific metric type for one file,
-     * updating both the checkpoint and activity data accordingly.
-     * 
-     * @param metricName - The type of metric to calculate (e.g., 'fileSize', 'wordCount')
-     * @param file - The Obsidian TFile to calculate metrics for
-     * @param latestData - The current state of all activity heatmap data
-     * @param isFirstTimeUpdate - Whether this is a first-time update (i.e. no existing data.json file)
-     * @returns An object containing the updated checkpoint and activity data
-     */
-    async calculateMetricsForFile(
-        metricName: MetricType,
-        file: TFile,
-        latestData: ActivityHeatmapData,
-        isFirstTimeUpdate: boolean
-    ): Promise<{ checkpoint: CheckpointData; activity: ActivityData }> {
-        const calculator = this.metricCalculators[metricName];
-        if (!calculator) {
-            throw new Error(`No calculator found for metric: ${metricName}`);
-        }
-
-        const checkpoint: CheckpointData = {};
-        const activity: ActivityData = { ...latestData.activityOverTime[metricName] };
-        
-        try {
-            const metricValue = await calculator(file);
-            checkpoint[file.path] = {
-                value: metricValue,
-                mtime: file.stat.mtime
-            };
-
-            if (!isFirstTimeUpdate) {
-                let absoluteDifference = metricValue;
-                if (latestData.checkpoints[metricName] && file.path in latestData.checkpoints[metricName]) {
-                    const previousRecord = latestData.checkpoints[metricName][file.path];
-                    absoluteDifference = calculateAbsoluteDifference(metricValue, previousRecord.value);
-                }
-                activity[getDateStringFromTimestamp(file.stat.mtime)] = (activity[getDateStringFromTimestamp(file.stat.mtime)] || 0) + absoluteDifference;
-            }
-        } catch (error) {
-            console.error(`Error calculating ${metricName} for file ${file.path}:`, error);
-        }
-
-        return { checkpoint, activity };
-    }
+   
 }
