+++ src/types/settings.ts
@@ -1,0 +1,7 @@
+import type { MetricType } from './metrics';
+
+export interface ActivityHeatmapSettings {
+    metricType: MetricType;
+    useMockData?: boolean;
+    year: string;
+}
